+++ app/src/main/java/woowacourse/movie/presenter/movieReservationResult/MovieReservationResultPresenter.kt
@@ -1,6 +1,5 @@
-package woowacourse.movie.presenter
+package woowacourse.movie.presenter.movieReservationResult
 
-import woowacourse.movie.MovieReservationResultContract
 import woowacourse.movie.view.model.TicketUiModel
 import woowacourse.movie.view.movieReservationResult.MovieReservationResultActivity
 import woowacourse.movie.view.movieReservationResult.MovieReservationResultActivity.Companion.KEY_TICKET
